Removing useless Helper class Moving some codes from Helper into TwigExtension
